feat: Custom assertion macro
SDL has a good assertion macro that I was using originally, but I've
wanted to be able to include a message - especially one that is
formatted. Therefore, I've made my own assertion macro based on
SDL_assert.

It's not nearly as robust as SDL's assert - it's not thread safe,
and has no protection against recursive assertions. But, it'll do for
now!