feat: Silent assertions + general improvements
Silent assertions will not show a message box, but are still recorded
in the assertion report, and are still logged to the console.

Assertion message boxes are now aware of the application window, if it
exists. This means the message box will stop the user focusing the
game window, which means they have to see the assertion.

Moved the assertion data struct to above the condition check, meaning
we can check for AlwaysIgnore before checking the condition, which
might be helpful if it's a particularly heavy condition. However,
perhaps loading the memory of the AssertionData every time we check it
instead of only when we fail it is a worse performance loss than
always checking ignored assertions. Besides, how mant assertions are
ignored? I might roll this part back.