fix: Fix semaphores and resizing
Released I had the wait and signal semaphore the wrong way around in
our submit info.. oops. Wondering how this ever worked on the other PC?
Still need to learn more about semaphores, how they're being used here
still confuses me a little.

Also, resizing was crashing because we destroy and recreate the draw
image when resizing. This means the image in the descriptors becomes
invalid, and we never set it again. For now, I've added a bodged fix
where we rewrite the image pointer when we resize.