feat: Preprocessor improvements
The preprocessor now creates a different last process time cache
for each config/platform combination. The main reason for this
is such that all the content will be copied over to each build
directory. Furthermore, in future, some content may compile
differently depending on the platform.

To accomodate that, the platform, config, and in/out dirs are now
public static strings. This means any processor will be able
to access them, and we don't need to pass strings around in the
main file.

Also, fixed issues with timezones with the last asset processing
time. The times were being loaded as local time, then had
the UTC -> local time conversion applied anyways, and then saved as
local time. This caused the times to jump by the local timezone's offset
from UTC every time the times were saved. I only noticed this because
I was previously in GMT, and we just switched over to BST!