feat: GPU selection and more
The renderer now has an option for the GPU index to use. The renderer
also exposes the names of the GPUs as an array, which we can use
to provide the user with a UI to select the GPU to use. The changes
only apply after restarting the app (although restarting in-app works.)

If no override is provided, the renderer will now select the GPU with
the highest local VRAM.

Renderer member variables are now properly initialised.

The app restart flag is now a static member of Application, instead of
a global variable in Main.

Moved fmt type formatters to their own header that is included at the
end of the PCH, such that we can use all types declared in the PCH
in the formatter definitions.

The log is now initialised by Application, making the Main file cleaner.

ApplicationSpec now has an author, which is used when getting the log
path.

SemVer packing no longer has empty bits in the middle, and instead
the most significant bits are empty. Seems more appropriate.